refactor(hooks): extract buildBottomAction from useRoomActions
