+++ src/screens/RoomScreen/hooks/useRoomActions.ts
@@ -22,6 +22,8 @@ import { useCallback } from 'react';
 import type { ActionIntent } from '@/screens/RoomScreen/policy/types';
 import type { LocalGameState } from '@/types/GameStateTypes';
 
+import { type BottomActionVM, buildBottomAction } from './bottomActionBuilder';
+
 // Re-export for architectural boundary contract (see hooks.boundary.test.ts)
 export type { ActionIntent };
 
@@ -90,17 +92,6 @@ interface UseRoomActionsResult {
 
   /** UI-only: schema-driven bottom action button view-model (visibility + label). */
   getBottomAction: () => BottomActionVM;
-}
-
-interface BottomActionVM {
-  buttons: BottomButton[];
-}
-
-interface BottomButton {
-  /** Stable key (align to schema step keys when possible). */
-  key: string; // 'save' | 'skip' | 'wolfEmpty' ...
-  label: string;
-  intent: ActionIntent;
 }
 
 // ─────────────────────────────────────────────────────────────────────────────
@@ -360,255 +351,33 @@ export function useRoomActions(gameContext: GameContext, deps: ActionDeps): UseR
   // UI-only: schema-driven bottom action button (skip / wolf empty vote / blocked)
   // ─────────────────────────────────────────────────────────────────────────
 
-  const getBottomAction = useCallback((): BottomActionVM => {
-    if (!gameState) return { buttons: [] };
-    if (roomStatus !== GameStatus.Ongoing) return { buttons: [] };
-    if (isAudioPlaying) return { buttons: [] };
-    if (!currentSchema) return { buttons: [] };
-
-    // All schemas (including groupConfirm): only visible to the actioner.
-    // For groupConfirm, determineActionerState already sets imActioner=true for all seated players.
-    if (!imActioner) return { buttons: [] };
-
-    // ─────────────────────────────────────────────────────────────────────────
-    // UI Hint（服务端广播驱动，UI 只读展示）
-    // 如果服务端广播了 currentActorHint，直接按 hint 渲染按钮，不再走 schema 分支。
-    // 使用 targetRoleIds 过滤：只有 actorRole 在 targetRoleIds 中才显示 hint。
-    // ─────────────────────────────────────────────────────────────────────────
-    const hint = gameState.ui?.currentActorHint;
-    const hintApplies = hint && actorRole && hint.targetRoleIds.includes(actorRole);
-    if (hintApplies) {
-      if (hint.bottomAction === 'skipOnly') {
-        return {
-          buttons: [
-            {
-              key: 'skip',
-              label: hint.message,
-              intent: { type: 'skip', targetSeat: -1, message: hint.message },
-            },
-          ],
-        };
-      }
-      if (hint.bottomAction === 'wolfEmptyOnly') {
-        return {
-          buttons: [
-            {
-              key: 'wolfEmpty',
-              label: hint.message,
-              intent: { type: 'wolfVote', targetSeat: -1, wolfSeat: actorSeatNumber ?? undefined },
-            },
-          ],
-        };
-      }
-      // hint 存在但没有 bottomAction 指示 → 按正常 schema 处理
-    }
-
-    // wolfRobot learned hunter gate: must view status before continuing
-    // Condition: learned hunter (regardless of canShootAsHunter true/false)
-    // This gate takes precedence over other bottom actions when wolfRobot step is active
-    if (
-      currentSchema.id === 'wolfRobotLearn' &&
-      gameState.wolfRobotReveal?.learnedRoleId === 'hunter' &&
-      gameState.wolfRobotHunterStatusViewed === false
-    ) {
-      // Schema-driven: read button text from schema (fail-fast if missing)
-      const gateButtonText = currentSchema.ui?.hunterGateButtonText;
-      if (!gateButtonText) {
-        throw new Error(
-          '[useRoomActions] wolfRobotLearn schema missing ui.hunterGateButtonText - schema-driven UI requires this field',
-        );
-      }
-      return {
-        buttons: [
-          {
-            key: 'viewHunterStatus',
-            label: gateButtonText,
-            intent: { type: 'wolfRobotViewHunterStatus', targetSeat: -1 },
-          },
-        ],
-      };
-    }
-
-    // wolfVote: show empty vote + cancel button when already voted
-    if (currentSchema.kind === 'wolfVote') {
-      const buttons: BottomButton[] = [];
-      const voted = actorSeatNumber !== null && hasWolfVoted(actorSeatNumber);
-
-      // Cancel vote button (withdraw = -2)
-      if (voted) {
-        buttons.push({
-          key: 'wolfCancel',
-          label: '取消投票',
-          intent: {
-            type: 'wolfVote',
-            targetSeat: -2,
-            wolfSeat: actorSeatNumber ?? undefined,
-          },
-        });
-      }
-
-      // Empty vote button (always available)
-      buttons.push({
-        key: 'wolfEmpty',
-        label: currentSchema.ui!.emptyVoteText!,
-        intent: {
-          type: 'wolfVote',
-          targetSeat: -1,
-          wolfSeat: actorSeatNumber ?? undefined,
-        },
-      });
-
-      return { buttons };
-    }
-
-    // chooseSeat/swap: honor canSkip
-    // NOTE: witchSave/witchPoison are chooseSeat sub-steps and should allow bottom skip.
-    if (currentSchema.kind === 'chooseSeat' || currentSchema.kind === 'swap') {
-      if (!currentSchema.canSkip) return { buttons: [] };
-      return {
-        buttons: [
-          {
-            key: 'skip',
-            label: currentSchema.ui!.bottomActionText!,
-            intent: {
-              type: 'skip',
-              targetSeat: -1,
-              message: currentSchema.ui!.bottomActionText!,
-            },
-          },
-        ],
-      };
-    }
-
-    // compound (witchAction): return two buttons (save + skip)
-    // Schema-driven: save step is 'confirmTarget' (fixed target from WITCH_CONTEXT),
-    //                poison step is 'chooseSeat' (user taps seat).
-    // NOTE: Sensitive info is from WitchContextPayload only.
-    if (
-      currentSchema.kind === 'compound' &&
-      currentSchema.id === 'witchAction' &&
-      currentSchema.steps?.length
-    ) {
-      const witchCtx = getWitchContext();
-      if (!witchCtx) return { buttons: [] };
-
-      // Schema-driven: save is confirmTarget (target = killedSeat), poison is chooseSeat
-      const saveStep = currentSchema.steps.find(
-        (s) => s.key === 'save' && s.kind === 'confirmTarget',
-      );
-      const poisonStep = currentSchema.steps.find(
-        (s) => s.key === 'poison' && s.kind === 'chooseSeat',
-      );
-
-      const buttons: BottomButton[] = [];
-
-      // 1) Save button (confirmTarget): only show when kill exists and canSave.
-      //    Target is fixed (witchCtx.killedSeat), user only confirms.
-      if (saveStep && witchCtx.killedSeat >= 0 && witchCtx.canSave) {
-        const label = `对${witchCtx.killedSeat + 1}号用解药`;
-        buttons.push({
-          key: 'save',
-          label,
-          intent: {
-            type: 'actionConfirm',
-            targetSeat: witchCtx.killedSeat,
-            message: saveStep.ui?.confirmText,
-            stepKey: 'save',
-          },
-        });
-      }
-
-      // 2) Skip button: always available; should mean save=false AND poison=false.
-      // We route through RoomScreen with stepKey='skipAll' (not a schema step) to avoid dual-submit.
-      // RoomScreen will translate this to extra {save:false, poison:false}.
-      const skipLabel = poisonStep!.ui!.bottomActionText!;
-      buttons.push({
-        key: 'skip',
-        label: skipLabel,
-        intent: { type: 'skip', targetSeat: -1, message: skipLabel, stepKey: 'skipAll' },
-      });
-
-      return { buttons };
-    }
-
-    // confirm schema (hunterConfirm/darkWolfKingConfirm)
-    // NOTE: Blocked case is already handled above (unified blocked check).
-    // Here we only handle the normal (not blocked) case: show confirm button.
-    if (currentSchema.kind === 'confirm') {
-      return {
-        buttons: [
-          {
-            key: 'confirm',
-            label: currentSchema.ui!.bottomActionText!,
-            intent: { type: 'confirmTrigger', targetSeat: -1 },
-          },
-        ],
-      };
-    }
-
-    // groupConfirm schema (piperHypnotizedReveal)
-    // Already acked → hide button; otherwise show "催眠状态" button.
-    if (currentSchema.kind === 'groupConfirm') {
-      const acks = gameState.piperRevealAcks ?? [];
-      if (actorSeatNumber !== null && acks.includes(actorSeatNumber)) {
-        return { buttons: [] };
-      }
-      return {
-        buttons: [
-          {
-            key: 'groupConfirmAck',
-            label: currentSchema.ui?.bottomActionText ?? '催眠状态',
-            intent: { type: 'groupConfirmAck', targetSeat: -1 },
-          },
-        ],
-      };
-    }
-
-    // multiChooseSeat (piperHypnotize): confirm + skip buttons
-    if (currentSchema.kind === 'multiChooseSeat') {
-      const buttons: BottomButton[] = [];
-      const count = multiSelectedSeats.length;
-
-      // Confirm button (only when at least 1 target selected)
-      if (count > 0) {
-        const rawLabel = currentSchema.ui?.confirmButtonText ?? '确认({count}人)';
-        buttons.push({
-          key: 'multiConfirm',
-          label: rawLabel.replace('{count}', String(count)),
-          intent: {
-            type: 'multiSelectConfirm',
-            targetSeat: -1,
-            targets: multiSelectedSeats,
-          },
-        });
-      }
-
-      // Skip button (if canSkip)
-      if (currentSchema.canSkip && currentSchema.ui?.bottomActionText) {
-        buttons.push({
-          key: 'skip',
-          label: currentSchema.ui.bottomActionText,
-          intent: { type: 'skip', targetSeat: -1, message: currentSchema.ui.bottomActionText },
-        });
-      }
-
-      return { buttons };
-    }
-
-    // skip: no generic bottom action
-    return { buttons: [] };
-  }, [
-    gameState,
-    getWitchContext,
-    imActioner,
-    isAudioPlaying,
-    currentSchema,
-    roomStatus,
-    actorSeatNumber,
-    actorRole,
-    hasWolfVoted,
-    multiSelectedSeats,
-  ]);
+  const getBottomAction = useCallback(
+    (): BottomActionVM =>
+      buildBottomAction({
+        gameState,
+        roomStatus,
+        isAudioPlaying,
+        currentSchema,
+        imActioner,
+        actorSeatNumber,
+        actorRole,
+        multiSelectedSeats,
+        hasWolfVoted,
+        getWitchContext,
+      }),
+    [
+      gameState,
+      getWitchContext,
+      imActioner,
+      isAudioPlaying,
+      currentSchema,
+      roomStatus,
+      actorSeatNumber,
+      actorRole,
+      hasWolfVoted,
+      multiSelectedSeats,
+    ],
+  );
 
   // ─────────────────────────────────────────────────────────────────────────
   // Auto-trigger intent (for roles that popup on turn start)
